[#361] Rename util to string.js & move trim.js to string.js
